MIPS Recompiler: Update it to senquack's part4 jit
We were using part 3 of the MIPS recompiler by senquack.
The reason why we could not use part 4 was in fact due to a GCC bug
incorrectly parsing save states.
It seems that my buildroot with my additional GCC patches actually fixed it...
That or it was GCC 8.2 maybe.

Since part4 now works, let's switch to that. Btw, he's about to release part5
soonish too.